move logic and load next events
